+++ tests/pystackql_tests.py
@@ -1,51 +1,6 @@
-import json, os, sys, platform, re, time, unittest, subprocess
-import pandas as pd
-from termcolor import colored
-import unittest.mock as mock
-
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 from pystackql import StackQL
 from pystackql.stackql_magic import StackqlMagic, load_ipython_extension
-
-server_port = 5466
-
-expected_properties = [
-    "bin_path", "download_dir", "package_version", "params", 
-    "parse_json", "platform", "server_mode", "sha", "version"
-]
-
-expected_version_pattern = r'^v?(\d+\.\d+\.\d+)$'  # should be 'vN.N.N' or 'N.N.N'
-
-expected_platform_pattern = r'^(Windows|Linux|Darwin) (\w+) \(([^)]+)\), Python (\d+\.\d+\.\d+)$'
-
-custom_windows_download_dir = 'C:\\temp'
-custom_mac_linux_download_dir = '/tmp'
-
-registry_pull_google_query = "REGISTRY PULL google"
-registry_pull_aws_query = "REGISTRY PULL aws"
-
-def registry_pull_resp_pattern(provider):
-    return r"%s provider, version 'v\d+\.\d+\.\d+' successfully installed" % provider
-
-google_query = f"""
-SELECT status, count(*) as num_instances
-FROM google.compute.instances
-WHERE project = '{os.environ['GCP_PROJECT']}' 
-AND zone = '{os.environ['GCP_ZONE']}'
-GROUP BY status
-"""
-
-regions = os.environ.get('REGIONS').split(',')
-
-async_queries = [
-    f"""
-    SELECT '{region}' as region, instanceType, COUNT(*) as num_instances
-    FROM aws.ec2.instances
-    WHERE region = '{region}'
-    GROUP BY instanceType
-    """
-    for region in regions
-]
+from .test_params import *
 
 def pystackql_test_setup(func):
     def wrapper(self):
@@ -56,24 +11,6 @@ def pystackql_test_setup(func):
         self.stackql = StackQL()
         func(self)
     return wrapper
-
-server_mode_header = colored("[SERVER MODE]", 'yellow')
-ipython_mode_header = colored("[MAGIC EXT]", 'blue')
-
-def print_test_result(test_name, condition, server_mode=False, is_ipython=False):
-    status_header = colored("[PASSED] ", 'green') if condition else colored("[FAILED] ", 'red')
-    headers = [status_header]
-
-    if server_mode:
-        headers.append(server_mode_header)
-    
-    if is_ipython:
-        headers.append(ipython_mode_header)
-
-    headers.append(test_name)
-
-    message = " ".join(headers)
-    print("\n" + message)
 
 class PyStackQLTestsBase(unittest.TestCase):
     pass